Reuse convert function
